% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{FebuxostatCVRisk-package}
\alias{FebuxostatCVRisk}
\alias{FebuxostatCVRisk-package}
\title{FebuxostatCVRisk: A Package for Comparative safety of Febuxostat on Cardiovascular Events and Death Risks Compared with Allopurinol}
\description{
A skeleton package, to confirm the adverse cardiovascular events and death risks of febuxostat compared with allopurinol in patients with gout.
}
\keyword{internal}
